Update AboutBox, file headers
